ARGO-3987 Make AMS related code common in flink jobs
