+++ flink_jobs_v2/AmsPublisher/src/main/java/ams/publisher/ArgoMessagingSource.java
@@ -1,4 +1,5 @@
-package argo.streaming;
+package ams.publisher;
+
 
 import java.security.KeyManagementException;
 import java.security.KeyStoreException;
