+++ flink_jobs_v2/ams_ingest_metric/src/main/java/argo/streaming/AmsIngestMetric.java
@@ -1,6 +1,7 @@
 package argo.streaming;
 
 
+import ams.publisher.ArgoMessagingSource;
 import java.util.Arrays;
 import java.util.concurrent.TimeUnit;
