+++ flink_jobs_v2/ams_ingest_sync/src/main/java/argo/streaming/AmsIngestSync.java
@@ -1,5 +1,6 @@
 package argo.streaming;
 
+import ams.publisher.ArgoMessagingSource;
 import java.util.concurrent.TimeUnit;
 
 import org.apache.flink.api.common.restartstrategy.RestartStrategies;
