+++ flink_jobs_v2/pom.xml
@@ -17,7 +17,7 @@
         <module>ProfilesManager</module>
         <module>batch_multi</module>
         <module>stream_status</module>
-         <module>AmsPublisher</module>         
+         <module>ams-connector</module>         
          <module>ams_ingest_metric</module>         
          <module>ams_ingest_sync</module>
