+++ flink_jobs_v2/stream_status/pom.xml
@@ -168,7 +168,12 @@ language governing permissions and limitations under the License. -->
             <artifactId>joda-time</artifactId>
             <version>1.6</version>
         </dependency>
-        
+           <dependency>
+            <groupId>argo.streaming</groupId>
+            <artifactId>AmsPublisher</artifactId>
+            <version>2.0.0</version>
+            <type>jar</type>
+        </dependency>     
     </dependencies>
 
     <profiles>
